fix: 컴포넌트별로 파일 나누기(타입 파일도 추가)
