+++ src/components/Login/types.ts
@@ -1,0 +1,17 @@
+// image
+export interface imageProps {
+    width: number;
+    height: number;
+    position: string;
+}
+
+// text
+export interface textProps {
+    innerText: string;
+}
+
+// style
+export interface NewCardProps {
+    padding: string;
+    margin: string;
+}
